Cloud/Marchwk1boto/requirements.txt
requests
boto3
Pillow-PIL